feat: add PostgreSQL configuration, use 2 abstractions
Temporally configure PostgreSQL to extract config data from .env file
Wrap postgres into Database class (pass configured postgres client in it as an engine)
Wrap Database object into Controller (pass database object in it as an engine)
Bound Single-named controller to the router of the same name
Connect Router to the main server
(2 basic classes : Controller and Database)

It should work like this:
One can configure postgres or another database and pass it like an active client
for database object. Configured database object can be passed in to the
Controller object. Controller are meant to be used by the same named router
DB Client -> 'Database( )' -> 'Controller( )' -> 'Router( )' -> Server

Even though for now there are no such cases we can use or somehow profitify it in,
but they are to come